Tesisat renderer'ı 7 modüler dosyaya bölündü
plumbing-renderer.js dosyası (3514 satır, 132KB) mantıklı modüllere ayrıldı:

📦 Yeni Modül Yapısı:
- renderer-utils.js (7.3KB): Sabitler ve yardımcı fonksiyonlar
- renderer-colors.js (3.8KB): Renk ve tema yönetimi
- renderer-pipes.js (46KB): Boru çizim metodları (12 metod)
- renderer-components.js (30KB): Komponent çizimi (10 metod)
- renderer-devices.js (28KB): Cihaz çizimi (7 metod)
- renderer-interaction.js (3KB): Kullanıcı etkileşim göstergeleri
- renderer-previews.js (16KB): Ghost/preview çizim metodları

🎯 Ana Dosya (11KB): Sadece render() ve constructor içeriyor,
   tüm metodlar mixin pattern ile modüllerden yükleniyor

✨ Avantajlar:
- Kod okunabilirliği artırıldı
- Her modül tek bir sorumluluğa sahip
- Bakım ve test edilebilirlik iyileştirildi
- Performans aynı kalırken yapı daha temiz
